Moved all the default handlers - they're all streaming receiver (apparently).
